add validate store

+++ app/Http/Controllers/CountryController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Country;
+use Dotenv\Validator;
 
 class CountryController extends Controller
 {
@@ -12,20 +13,40 @@ class CountryController extends Controller
     }
 
     public function show($id) {
-        return response()->json(Country::find($id), 200);
+        $country = Country::find($id);
+        if (is_null($country)) {
+            return response()->json(['error' => true, 'message' => 'Not found'], 404);
+        }
+        return response()->json($country, 200);
     }
 
     public function store(Request $request) {
-        $country = Country::create($request->all());
+        $data = $this->validate(
+            $request, [
+                'alias' => 'required',
+                'name' => 'required|min:3',
+                'name_en' => 'required|min:3'
+            ]
+        );
+        $country = new Country();
+        $country->fill($data)->save();
         return response()->json($country, 201);
     }
 
-    public function update(Request $request, Country $country) {
+    public function update(Request $request, $id) {
+        $country = Country::find($id);
+        if (is_null($country)) {
+            return response()->json(['error' => true, 'message' => 'Not found'], 404);
+        }
         $country->update($request->all());
         return response()->json($country, 200);
     }
 
-    public function destroy(Request $request, Country $country) {
+    public function destroy(Request $request, $id) {
+        $country = Country::find($id);
+        if (is_null($country)) {
+            return response()->json(['error' => true, 'message' => 'Not found'], 404);
+        }
         $country->delete();
         return response()->json('Country is deleted', 200);
     }
